+++ app/Http/Controllers/dashboard/admin/StoreCategoryController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\dashboard\admin;
 
+use App\Enums\PermissionsEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Dashboard\StoreCategoryRequest;
 use App\Http\Resources\StoreCategoryResource;
@@ -13,6 +14,8 @@ class StoreCategoryController extends Controller
 {
     public function index(Request $request)
     {
+        abort_unless($request->user('admin')->can(PermissionsEnum::STORE_CATEGORIES_INDEX->value), 403);
+
         $categories = StoreCategory::query()
             ->search($request->get('tableSearch'))
             ->orderBy($request->get('sort', 'id'), $request->get('direction', 'desc'))
@@ -26,6 +29,8 @@ class StoreCategoryController extends Controller
 
     public function create()
     {
+        abort_unless(request()->user('admin')->can(PermissionsEnum::STORE_CATEGORIES_CREATE->value), 403);
+
         return Inertia::render('admin/store-categories/create', [
             'category' => StoreCategoryResource::make(new StoreCategory())->serializeForForm(),
         ]);
@@ -33,12 +38,16 @@ class StoreCategoryController extends Controller
 
     public function store(StoreCategoryRequest $request)
     {
+        abort_unless($request->user('admin')->can(PermissionsEnum::STORE_CATEGORIES_CREATE->value), 403);
+
         StoreCategory::create($request->validated());
         return to_route('admin.store-categories.index')->with('success', __('messages.created_successfully'));
     }
 
     public function edit($id)
     {
+        abort_unless(request()->user('admin')->can(PermissionsEnum::STORE_CATEGORIES_UPDATE->value), 403);
+
         $category = StoreCategory::findOrFail($id);
         return Inertia::render('admin/store-categories/edit', [
             'category' => StoreCategoryResource::make($category)->serializeForForm(),
@@ -47,6 +56,8 @@ class StoreCategoryController extends Controller
 
     public function update($id, StoreCategoryRequest $request)
     {
+        abort_unless($request->user('admin')->can(PermissionsEnum::STORE_CATEGORIES_UPDATE->value), 403);
+
         $category = StoreCategory::findOrFail($id);
         $category->update($request->validated());
 
@@ -57,6 +68,8 @@ class StoreCategoryController extends Controller
 
     public function destroy($id)
     {
+        abort_unless(request()->user('admin')->can(PermissionsEnum::STORE_CATEGORIES_DESTROY->value), 403);
+
         StoreCategory::destroy($id);
 
         return redirect()
